Bugs: In empty Dropdown popup - "Add new customer" is coming... Fix customer max width of drop down multiple errors are showing in notification ... Remove horizontal bar from pages... Done Invoive number from Invoice id... Footer should always come at the end of page.

+++ src/main/java/org/baeldung/persistence/dao/IInvoiceRespository.java
@@ -13,7 +13,7 @@ import org.springframework.data.jpa.repository.Query;
  * @author sachinbabujettipraksh
  *
  */
-public interface IInvoiceRespository extends JpaRepository<Invoice, Long> {
+public interface IInvoiceRespository extends JpaRepository<Invoice, Integer> {
 
     @Override
     void delete(Invoice invoice);
